refac: extracting helpers from e2e tests
