+++ e2e/test/follower/e2e-pagination.test.ts
@@ -1,45 +1,26 @@
-import axios from "axios";
 import { expect } from "chai";
 
 import { ALICE } from "../helpers/account";
 import { CHAIN_ID_DEC, send, sendAndGetFullResponse } from "../helpers/rpc";
+import { FOLLOWER_URL, rpcCall, waitForFollowerBlock, waitForReceipt } from "./helpers";
 
-// This test only makes sense with a leader and a follower running with small response size limits:
-// the leader with MAX_RESPONSE_SIZE_BYTES and the follower with EXTERNAL_RPC_MAX_RESPONSE_SIZE_BYTES.
-// It is executed by the `just e2e-leader-follower-pagination` recipe.
-//
-// What it covers, in order:
-// 1. Confirms a response that fits the limits is served normally, with no pagination
-//    envelope — this is what old followers receive (byte-identical legacy path).
-// 2. Mines a block whose serialized `stratus_getBlockAndReceipts` response exceeds the limits.
-// 3. Confirms the old single-parameter call fails with the oversized response error (-32008),
-//    which is what would stall an importer before pagination existed.
-// 4. Reassembles the response from the paginated envelope chunks and validates the content.
-// 5. Waits for the follower importer to sync the fat block, proving the importer paginated it.
+// Requires the `just e2e-leader-follower-pagination` recipe: leader and follower both running
+// with MAX_RESPONSE_SIZE_BYTES=8192. The rule under test is simple — when the serialized response
+// exceeds the limit it must be paginated, and the follower must still sync the block.
 
-// Fat transaction payload: enough data for the block + receipts response to exceed the limits.
+const MAX_RESPONSE_BYTES = 8192;
 const FAT_TX_DATA_BYTES = 50_000;
-
-// Small chunk budget to force several round trips during reassembly.
-const CHUNK_BUDGET = 1024;
-
-const FOLLOWER_URL = process.env.FOLLOWER_URL || "http://localhost:3001";
-const LEADER_MAX_RESPONSE_BYTES = parseInt(process.env.MAX_RESPONSE_SIZE_BYTES || "8192");
-const FOLLOWER_SYNC_TIMEOUT_MS = 120_000;
-const POLL_INTERVAL_MS = 500;
 
 describe("Pagination", () => {
     it("paginates oversized importer responses and keeps the follower syncing", async () => {
-        // compatibility: a small early block is served normally, without any pagination envelope,
-        // so an old follower (or any client) keeps receiving the exact same responses as before
+        // a fitting response is served normally, with no envelope, so old followers are unaffected
         const earlyBlock = await send("eth_getBlockByNumber", ["0x1", false]);
         expect(earlyBlock).to.not.be.null;
         const small = await send("stratus_getBlockAndReceipts", [earlyBlock.hash]);
         expect(small.__stratus_paginated__).to.be.undefined;
         expect(small.block.number).to.equal("0x1");
 
-        // send a fat contract-deployment transaction and wait for it to be mined;
-        // the deployed code always fails, but the fat transaction data still makes the block heavy
+        // fat contract deployment: the code always fails, but the fat data makes the response oversized
         const nonce = await send("eth_getTransactionCount", [ALICE.address]);
         const signedTx = await ALICE.signer().signTransaction({
             data: "0x" + "ab".repeat(FAT_TX_DATA_BYTES),
@@ -52,24 +33,21 @@ describe("Pagination", () => {
 
         const receipt = await waitForReceipt(txHash);
         const fatBlockNumber = parseInt(receipt.blockNumber, 16);
+        const fatBlockHash = receipt.blockHash;
 
-        // the block hash fits in one response (hashes only), but the full response does not
-        const fatBlock = await send("eth_getBlockByNumber", [toHexNumber(fatBlockNumber), false]);
-        const fatBlockHash = fatBlock.hash;
-
-        // old-style single-parameter call: the oversized response is rejected by the server,
-        // which is exactly what would get an old importer stuck retrying forever
+        // the old single-parameter call fails with the oversized response error (-32008),
+        // which is exactly what would stall an importer before pagination existed
         const legacy = await sendAndGetFullResponse("stratus_getBlockAndReceipts", [fatBlockHash]);
         expect(legacy.data.error).to.not.be.undefined;
         expect(legacy.data.error.code).to.equal(-32008);
 
-        // paginated reassembly
+        // paginated reassembly, with a small chunk budget to force several round trips
         let assembled = "";
         let total = 0;
         for (let offset = 0; total === 0 || assembled.length < total; offset = assembled.length) {
             const envelope = await send("stratus_getBlockAndReceipts", [
                 fatBlockHash,
-                { offset: offset, chunk_budget: CHUNK_BUDGET },
+                { offset: offset, chunk_budget: 1024 },
             ]);
             expect(envelope.__stratus_paginated__).to.not.be.undefined;
             total = envelope.__stratus_paginated__.total;
@@ -77,9 +55,9 @@ describe("Pagination", () => {
             assembled += envelope.__stratus_paginated__.chunk;
         }
         expect(assembled.length).to.equal(total);
-        expect(total).to.be.greaterThan(LEADER_MAX_RESPONSE_BYTES, "the block response should be oversized");
+        expect(total).to.be.greaterThan(MAX_RESPONSE_BYTES, "the block response should be oversized");
 
-        // reassembled content matches the actual block
+        // the reassembled content matches the block
         const response = JSON.parse(assembled);
         expect(response.block.hash).to.equal(fatBlockHash);
         expect(parseInt(response.block.number, 16)).to.equal(fatBlockNumber);
@@ -94,48 +72,3 @@ describe("Pagination", () => {
         expect(followerReceipt.result.blockNumber).to.equal(receipt.blockNumber);
     });
 });
-
-// Sends a JSON-RPC request to an arbitrary node, tolerating error statuses.
-async function rpcCall(url: string, method: string, params: any[] = []): Promise<any> {
-    const response = await axios.post(
-        url,
-        { jsonrpc: "2.0", id: 1, method: method, params: params },
-        { validateStatus: () => true },
-    );
-    return response.data;
-}
-
-// Polls the leader until the transaction is mined and returns its receipt.
-async function waitForReceipt(txHash: string): Promise<any> {
-    const deadline = Date.now() + 60_000;
-    while (Date.now() < deadline) {
-        const receipt = await send("eth_getTransactionReceipt", [txHash]);
-        if (receipt) {
-            return receipt;
-        }
-        await sleep(POLL_INTERVAL_MS);
-    }
-    throw new Error(`transaction ${txHash} was not mined in time`);
-}
-
-// Polls the follower until it reaches the target block number.
-async function waitForFollowerBlock(target: number): Promise<void> {
-    const deadline = Date.now() + FOLLOWER_SYNC_TIMEOUT_MS;
-    while (Date.now() < deadline) {
-        const body = await rpcCall(FOLLOWER_URL, "eth_blockNumber", []);
-        const current = parseInt(body.result, 16);
-        if (current >= target) {
-            return;
-        }
-        await sleep(POLL_INTERVAL_MS);
-    }
-    throw new Error(`follower did not reach block ${target} in time`);
-}
-
-function sleep(ms: number): Promise<void> {
-    return new Promise((resolve) => setTimeout(resolve, ms));
-}
-
-function toHexNumber(value: number): string {
-    return "0x" + value.toString(16);
-}
